fix: the translation is now disabled on launch
As the '--user-data-dir' option is provided on the browsers launch with a dumb
directory, the user options are not kept (deliberately). Therefore if the user's
browser language is not English, it will ask to translate the Karma page at each
launch, which can be annoying. To address this issue, this fix manually disables
the translation by passing the appropriate command line argument.