feat(config): Allow --user-data-dir to be set in launcher properties
Introduce a new parameter to custom launder properties: `chromeUserDir`
which allows to set the `--user-data-dir` command line option to point
to a specific place. Rationale:
Sometimes it is useful to run Chrome under Karma with certain plugins
loaded. One use-case is to have the **JetBrains IDE Support** plugin to
attach the IntelliJ debugger. We can do it with `--load-extension` but
this can also be prohibited by corporate policy. So one way to solve it
is to have the `--user-data-dir` point to a project subdirectory and to
install the plugin manually on the first run. It is preserved of course
on subsequent runs.