chore(linter): switch from eslint to standard
We are currently using eslint with standard plugin for linting the
project's code. This is practically the same as using standard alone,
and has the advantage of not needing all the eslint packages, resulting
in a lot less setup and noise from greenkeeper.

Standard version was downgraded to v7.1.2 to maintain compatiblity with
supported node engines.