fix: remove npm install from travis
This install makes travis install a recent npm that does not support
node 0.10 nor 0.12, causing the build to fail for these versions of
node.

This commit removes the npm install from the travis config file to avoid
installing a recent npm, and use a version of npm that supports each
node version.